Update msmain.py
Added code to read in boards and return all values in a python array.

Because you can't enter characters in the array, 0s are for empty squares, 9s are for unknown squares, 7s are for bombs and 8s are for flags.

As of right now the program can completely read a board and convert the whole thing into a python array. From here, the goal is to use the information contained in the python array to make the correct moves.